refactor(node): split tasks.rs into submodules
Break the monolithic 2987-line tasks.rs into focused submodules under
src/tasks/ for improved maintainability and code navigation.
